+++ README.md
@@ -2,6 +2,12 @@
 
 > 领域事件的实践DEMO
 
+#### 更新日志：
+
+2017-08-29
+
+补充了最新的领域事件实现方式，以下单为例
+
 #### DEMO简介
 
   本DEMO以常用的用户注册\登录\登出为例,介绍领域事件的应用以及几种实现方式的不同.
@@ -41,7 +47,10 @@
 
     观察日志发现,用户登出监听器和request的处理在同一个线程中,说明Spring自带的是同步的.
 
-  设计初衷见我的博客-[实现领域事件](http://michael-j.net/2016/01/19/实现领域事件/)
+  设计初衷见我的博客:
+  [实现领域事件](http://michael-j.net/2016/01/19/实现领域事件/)
+  [再谈领域事件](http://michael-j.net/2017/08/13/再谈领域事件/)
+  
 
 #### MIT LICENSE
