--- src/main/java/com/haocai/EventLightApplication.java
@@ -1,12 +1,0 @@
-package com.haocai;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EventLightApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EventLightApplication.class, args);
-	}
-}
