--- src/main/java/com/haocai/core/event/DomainEvent.java
@@ -1,20 +1,0 @@
-package com.haocai.core.event;
-
-import java.util.Date;
-
-/**
- * Created by Michael Jiang on 16/1/12.
- */
-public abstract class DomainEvent {
-    private Date occurredTime;
-
-    protected abstract String identify();
-
-    public DomainEvent() {
-        occurredTime =new Date();
-    }
-
-    public Date getOccurredTime() {
-        return occurredTime;
-    }
-}
