--- src/main/java/com/haocai/core/event/DomainEventPublisher.java
@@ -1,15 +1,0 @@
-package com.haocai.core.event;
-
-/**
- * 领域事件发射器
- * Created by Michael Jiang on 16/1/12.
- */
-public interface DomainEventPublisher<T extends DomainEvent> {
-    String identify();
-
-    void register(Object listener);
-
-    void publish(T event);
-
-    void asyncPublish(T event);
-}
