--- src/main/java/com/haocai/core/user/event/LogoutEvent.java
@@ -1,27 +1,0 @@
-package com.haocai.core.user.event;
-
-import org.springframework.context.ApplicationEvent;
-
-/**
- * user logout event
- * Created by Michael Jiang on 16/1/13.
- */
-public class LogoutEvent extends ApplicationEvent {
-    private String name;
-
-    public LogoutEvent(Object source, String name) {
-        super(source);
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public String toString() {
-        return "LogoutEvent{" +
-                "name='" + name + '\'' +
-                '}';
-    }
-}
