--- src/main/java/com/haocai/core/user/event/UserEventType.java
@@ -1,9 +1,0 @@
-package com.haocai.core.user.event;
-
-/**
- * Created by Michael Jiang on 16/1/12.
- */
-public enum  UserEventType {
-    REGISTER,
-    LOGIN
-}
