+++ src/main/java/net/michael_j/ddd/EventLightApplication.java
@@ -1,0 +1,12 @@
+package net.michael_j.ddd;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EventLightApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(EventLightApplication.class, args);
+	}
+}
