+++ src/main/java/net/michael_j/ddd/core/user/event/UserEventType.java
@@ -1,0 +1,9 @@
+package net.michael_j.ddd.core.user.event;
+
+/**
+ * Created by Michael Jiang on 16/1/12.
+ */
+public enum  UserEventType {
+    REGISTER,
+    LOGIN
+}
