+++ src/main/java/net/michael_j/ddd/web/UserController.java
@@ -1,0 +1,35 @@
+package net.michael_j.ddd.web;
+
+import net.michael_j.ddd.application.service.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * Created by Michael Jiang on 16/1/12.
+ */
+@RestController
+@RequestMapping("/user")
+public class UserController {
+    @Autowired
+    private UserService userService;
+
+    @RequestMapping(value = "/register.do", method = RequestMethod.GET)
+    public String register(String name, Integer age) {
+        userService.register(name, age);
+        return "register completed!";
+    }
+
+    @RequestMapping(value = "/login.do", method = RequestMethod.GET)
+    public String login(String name) {
+        userService.login(name);
+        return "login completed!";
+    }
+
+    @RequestMapping(value = "/logout.do", method = RequestMethod.GET)
+    public String logout(String name) {
+        userService.logout(name);
+        return "logout completed!";
+    }
+}
